Rework view implementation.
Up till now, a view was a single type of entity, only supporting the
normal xdg toplevel surface views. With this rework, each view can have
its own implementation that implements the `bsi_view_impl` interface.
This enables us to have different types of views, for example an xwayland
view.
