+++ bonsai/desktop/decoration.c
@@ -21,18 +21,7 @@
 #include "bonsai/log.h"
 #include "bonsai/output.h"
 #include "bonsai/server.h"
-
-void
-decorations_add(struct bsi_server* server, struct bsi_xdg_decoration* deco)
-{
-    wl_list_insert(&server->scene.xdg_decorations, &deco->link_server);
-}
-
-void
-decorations_remove(struct bsi_xdg_decoration* deco)
-{
-    wl_list_remove(&deco->link_server);
-}
+#include "bonsai/util.h"
 
 struct bsi_xdg_decoration*
 decoration_init(struct bsi_xdg_decoration* deco,
@@ -43,10 +32,8 @@ decoration_init(struct bsi_xdg_decoration* deco,
     deco->server = server;
     deco->view = view;
     deco->xdg_decoration = wlr_deco;
-    deco->view->xdg_decoration_mode = wlr_deco->requested_mode;
-    deco->view->xdg_decoration = deco;
-    // const float color[4] = { 0.0f, 0.0f, 0.0f, 1.0f };
-    // deco->scene_rect = wlr_scene_rect_create(view->scene_node, 0, 0, color);
+    deco->view->decoration_mode = wlr_deco->requested_mode;
+    deco->view->decoration = deco;
     return deco;
 }
 
@@ -54,10 +41,11 @@ void
 decoration_update(struct bsi_xdg_decoration* decoration)
 {
     struct bsi_view* view = decoration->view;
-    bsi_debug("Update decoration for view '%s'", view->toplevel->app_id);
+    bsi_debug("Update decoration for view '%s'",
+              view->wlr_xdg_toplevel->app_id);
     wlr_scene_node_set_enabled(&decoration->scene_rect->node, view->mapped);
     wlr_scene_rect_set_size(
-        decoration->scene_rect, view->box.width + 4, view->box.height + 4);
+        decoration->scene_rect, view->geom.width + 4, view->geom.height + 4);
     wlr_scene_node_set_position(&decoration->scene_rect->node, -2, -2);
     wlr_scene_node_lower_to_bottom(&decoration->scene_rect->node);
     const float color_inactive[4] = { 0.2f, 0.2f, 0.2f, 1.0f };
@@ -91,7 +79,7 @@ decoration_iter(struct wlr_scene_buffer* buffer,
         return;
 
     struct wlr_box vb;
-    wlr_xdg_surface_get_geometry(view->toplevel->base, &vb);
+    wlr_xdg_surface_get_geometry(view->wlr_xdg_toplevel->base, &vb);
 
     cairo_surface_t* surface = cairo_image_surface_create(
         CAIRO_FORMAT_ARGB32, vb.width + 5, vb.height + 5);
@@ -106,9 +94,9 @@ decoration_iter(struct wlr_scene_buffer* buffer,
     cairo_select_font_face(
         cr, "SansSerif", CAIRO_FONT_SLANT_NORMAL, CAIRO_FONT_WEIGHT_NORMAL);
     cairo_set_font_size(cr, 1.5);
-    cairo_text_extents(cr, view->toplevel->app_id, &te);
+    cairo_text_extents(cr, view->wlr_xdg_toplevel->app_id, &te);
     cairo_move_to(cr, vb.width / 2.0 - te.width / 2.0, 2.0);
-    cairo_show_text(cr, view->toplevel->app_id);
+    cairo_show_text(cr, view->wlr_xdg_toplevel->app_id);
 
     size_t isize = cairo_image_surface_get_stride(surface) *
                    cairo_image_surface_get_height(surface);
@@ -131,11 +119,9 @@ decoration_iter(struct wlr_scene_buffer* buffer,
     cairo_destroy(cr);
 }
 
-/**
- * Handlers
- */
-void
-handle_xdg_decoration_destroy(struct wl_listener* listener, void* data)
+/* Handlers */
+static void
+handle_destroy(struct wl_listener* listener, void* data)
 {
     bsi_debug("Got event destroy from wlr_server_decoration");
     struct bsi_xdg_decoration* server_deco =
@@ -144,8 +130,8 @@ handle_xdg_decoration_destroy(struct wl_listener* listener, void* data)
     decoration_destroy(server_deco);
 }
 
-void
-handle_xdg_decoration_request_mode(struct wl_listener* listener, void* data)
+static void
+handle_request_mode(struct wl_listener* listener, void* data)
 {
     bsi_debug("Got event request_mode from wlr_server_decoration");
     struct bsi_xdg_decoration* server_deco =
@@ -154,9 +140,46 @@ handle_xdg_decoration_request_mode(struct wl_listener* listener, void* data)
     struct bsi_view* view = deco->surface->data;
 
     bsi_info("View with app-id '%s', requested SSD %d",
-             view->toplevel->app_id,
+             view->wlr_xdg_toplevel->app_id,
              deco->requested_mode ==
                  WLR_XDG_TOPLEVEL_DECORATION_V1_MODE_SERVER_SIDE);
 
     wlr_xdg_toplevel_decoration_v1_set_mode(deco, deco->requested_mode);
+}
+
+/* Global server handlers. */
+void
+handle_xdg_decoration_manager_new_decoration(struct wl_listener* listener,
+                                             void* data)
+{
+    bsi_debug("Got event new_decoration from wlr_decoration_manager");
+
+    struct bsi_server* server =
+        wl_container_of(listener, server, listen.new_decoration);
+    struct wlr_xdg_toplevel_decoration_v1* toplevel_deco = data;
+
+    struct bsi_view* view =
+        ((struct wlr_scene_node*)toplevel_deco->surface->data)->data;
+    assert(view);
+
+    struct bsi_xdg_decoration* xdg_deco =
+        calloc(1, sizeof(struct bsi_xdg_decoration));
+    decoration_init(xdg_deco, server, view, toplevel_deco);
+
+    util_slot_connect(&toplevel_deco->events.destroy,
+                      &xdg_deco->listen.destroy,
+                      handle_destroy);
+    util_slot_connect(&toplevel_deco->events.request_mode,
+                      &xdg_deco->listen.request_mode,
+                      handle_request_mode);
+
+    decorations_add(server, xdg_deco);
+
+    // bsi_decoration_draw(xdg_deco);
+
+    // TODO: Until server side decorations are implemented, refuse to set SSD.
+    wlr_xdg_toplevel_decoration_v1_set_mode(
+        toplevel_deco, WLR_XDG_TOPLEVEL_DECORATION_V1_MODE_CLIENT_SIDE);
+    // wlr_xdg_toplevel_decoration_v1_set_mode(toplevel_deco,
+    // toplevel_deco->requested_mode);
 }
