+++ include/bonsai/events.h
@@ -21,66 +21,22 @@ extern bsi_notify_func_t handle_pointer_grab_begin_notify;
 extern bsi_notify_func_t handle_pointer_grab_end_notify;
 extern bsi_notify_func_t handle_keyboard_grab_begin_notify;
 extern bsi_notify_func_t handle_keyboard_grab_end_notify;
-extern bsi_notify_func_t handle_touch_grab_begin_notify;
-extern bsi_notify_func_t handle_touch_grab_end_notify;
 extern bsi_notify_func_t handle_request_set_cursor_notify;
 extern bsi_notify_func_t handle_request_set_selection_notify;
 extern bsi_notify_func_t handle_request_set_primary_selection_notify;
-extern bsi_notify_func_t handle_request_start_drag_notify;
 /* wlr_xdg_shell */
-extern bsi_notify_func_t handle_xdgshell_new_surface;
+extern bsi_notify_func_t handle_xdg_shell_new_surface;
 /* wlr_layer_shell_v1 */
-extern bsi_notify_func_t handle_layershell_new_surface;
+extern bsi_notify_func_t handle_layer_shell_new_surface;
 /* wlr_xdg_decoration_manager_v1 */
-extern bsi_notify_func_t handle_xdg_deco_manager_new_decoration;
+extern bsi_notify_func_t handle_xdg_decoration_manager_new_decoration;
 /* wlr_xdg_activation_v1 */
 extern bsi_notify_func_t handle_xdg_request_activate;
 /* wlr_idle_inhibit_manager_v1 */
 extern bsi_notify_func_t handle_idle_manager_new_inhibitor;
 extern bsi_notify_func_t handle_idle_activity_notify;
 /* wlr_session_lock_manager_v1 */
-extern bsi_notify_func_t handle_session_new_lock;
-
-/*
- * bsi_view
- */
-extern bsi_notify_func_t handle_xdg_surf_destroy;
-extern bsi_notify_func_t handle_xdg_surf_map;
-extern bsi_notify_func_t handle_xdg_surf_unmap;
-extern bsi_notify_func_t handle_toplvl_request_maximize;
-extern bsi_notify_func_t handle_toplvl_request_fullscreen;
-extern bsi_notify_func_t handle_toplvl_request_minimize;
-extern bsi_notify_func_t handle_toplvl_request_move;
-extern bsi_notify_func_t handle_toplvl_request_resize;
-extern bsi_notify_func_t handle_toplvl_request_show_window_menu;
-
-/*
- * bsi_output
- */
-extern bsi_notify_func_t handle_output_frame;
-extern bsi_notify_func_t handle_output_destroy;
-extern bsi_notify_func_t handle_output_damage_frame;
-
-/*
- * bsi_input_{pointer,keyboard}
- */
-extern bsi_notify_func_t handle_pointer_motion;
-extern bsi_notify_func_t handle_pointer_motion_absolute;
-extern bsi_notify_func_t handle_pointer_button;
-extern bsi_notify_func_t handle_pointer_axis;
-extern bsi_notify_func_t handle_pointer_frame;
-extern bsi_notify_func_t handle_pointer_swipe_begin;
-extern bsi_notify_func_t handle_pointer_swipe_update;
-extern bsi_notify_func_t handle_pointer_swipe_end;
-extern bsi_notify_func_t handle_pointer_pinch_begin;
-extern bsi_notify_func_t handle_pointer_pinch_update;
-extern bsi_notify_func_t handle_pointer_pinch_end;
-extern bsi_notify_func_t handle_pointer_hold_begin;
-extern bsi_notify_func_t handle_pointer_hold_end;
-extern bsi_notify_func_t handle_keyboard_key;
-extern bsi_notify_func_t handle_keyboard_modifiers;
-/* bsi_input_{pointer,keyboard} -> wlr_input_device::events::destroy */
-extern bsi_notify_func_t handle_input_device_destroy;
+extern bsi_notify_func_t handle_session_lock_manager_new_lock;
 
 /*
  * bsi_workspace
@@ -88,61 +44,3 @@ extern bsi_notify_func_t handle_input_device_destroy;
 extern bsi_notify_func_t handle_server_workspace_active;
 extern bsi_notify_func_t handle_output_workspace_active;
 extern bsi_notify_func_t handle_view_workspace_active;
-
-/*
- * bsi_layer_surface_toplevel
- */
-/* wlr_layer_surface_v1 */
-extern bsi_notify_func_t handle_layershell_toplvl_map;
-extern bsi_notify_func_t handle_layershell_toplvl_unmap;
-extern bsi_notify_func_t handle_layershell_toplvl_destroy;
-extern bsi_notify_func_t handle_layershell_toplvl_new_popup;
-/* wlr_surface -> wlr_layer_surface::surface */
-extern bsi_notify_func_t handle_layershell_toplvl_commit;
-extern bsi_notify_func_t handle_layershell_toplvl_new_subsurface;
-
-/*
- * bsi_layer_surface_popup
- */
-/* wlr_xdg_surface -> wlr_xdg_popup::base */
-extern bsi_notify_func_t handle_layershell_popup_map;
-extern bsi_notify_func_t handle_layershell_popup_unmap;
-extern bsi_notify_func_t handle_layershell_popup_destroy;
-extern bsi_notify_func_t handle_layershell_popup_commit;
-extern bsi_notify_func_t handle_layershell_popup_new_popup;
-
-/*
- * bsi_layer_surface_subsurface
- */
-/* wlr_subsurface */
-extern bsi_notify_func_t handle_layershell_subsurface_map;
-extern bsi_notify_func_t handle_layershell_subsurface_unmap;
-extern bsi_notify_func_t handle_layershell_subsurface_destroy;
-extern bsi_notify_func_t handle_layershell_subsurface_commit;
-
-/*
- * bsi_server_decoration
- */
-extern bsi_notify_func_t handle_xdg_decoration_destroy;
-extern bsi_notify_func_t handle_xdg_decoration_request_mode;
-
-/*
- * bsi_idle_inhibitor
- */
-extern bsi_notify_func_t handle_idle_inhibitor_destroy;
-
-/*
- * bsi_session_lock
- */
-extern bsi_notify_func_t handle_session_lock_new_surface;
-extern bsi_notify_func_t handle_session_lock_unlock;
-extern bsi_notify_func_t handle_session_lock_destroy;
-
-/*
- * bsi_session_lock_surface
- */
-extern bsi_notify_func_t handle_lock_surface_map;
-extern bsi_notify_func_t handle_lock_surface_destroy;
-extern bsi_notify_func_t handle_lock_surface_surface_commit;
-extern bsi_notify_func_t handle_lock_surface_output_mode;
-extern bsi_notify_func_t handle_lock_surface_output_commit;
